Tweak input manipulation formats
This extracts the notion of “not consuming the input” from the input
manipulation formats, and adjusts the names of these formats to work
more nicely together.

This was partly inspired by the paper “Interval Parsing Grammars for
File Format Parsing” - except here we default to consuming bytes from
the input as opposed to parsing in-place.